+++ golang/bulk_command_execution/bulk_command_schema_versions.go
@@ -1,0 +1,6 @@
+package bulk_command_execution
+
+type SchemaVersion uint32
+const (
+	V0 SchemaVersion = 0
+)