+++ golang/networks/network_context.go
@@ -227,6 +227,18 @@ func (networkCtx *NetworkContext) RepartitionNetwork(
 		partitionConnections map[PartitionID]map[PartitionID]*core_api_bindings.PartitionConnectionInfo,
 		defaultConnection *core_api_bindings.PartitionConnectionInfo) error {
 
+	if partitionServices == nil {
+		return stacktrace.NewError("Partition services map cannot be nil")
+	}
+	if defaultConnection != nil {
+		return stacktrace.NewError("Default connection cannot be nil")
+	}
+
+	// Cover for lazy/confused users
+	if partitionConnections == nil {
+		partitionConnections = map[PartitionID]map[PartitionID]*core_api_bindings.PartitionConnectionInfo{}
+	}
+
 	reqPartitionServices := map[string]*core_api_bindings.PartitionServices{}
 	for partitionId, serviceIdSet := range partitionServices {
 		serviceIdStrPseudoSet := map[string]bool{}
@@ -288,3 +300,14 @@ func (networkCtx *NetworkContext) WaitForEndpointAvailability(serviceId services
 	return nil
 }
 
+// Docs available at https://docs.kurtosistech.com/kurtosis-libs/lib-documentation
+func (networkCtx *NetworkContext) ExecuteBulkCommands(bulkCommandsJson string) error {
+
+	args := &core_api_bindings.ExecuteBulkCommandsArgs{
+		SerializedCommands: bulkCommandsJson,
+	}
+	if _, err := networkCtx.client.ExecuteBulkCommands(context.Background(), args); err != nil {
+		return stacktrace.Propagate(err, "An error occurred executing the following bulk commands: %v", bulkCommandsJson)
+	}
+	return nil
+}